@@ -23,7 +23,7 @@
     "html-webpack-plugin": "^4.0.0",
     "react-hot-loader": "^4.12.20",
     "style-loader": "^1.1.3",
-    "webpack": "^4.42.0",
+    "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
     "webpack-dev-server": "^3.10.3",
     "webpack-merge": "^4.2.2"
@@ -40,6 +40,7 @@
     "react-router-dom": "^5.1.2",
     "redux": "^4.0.5",
     "redux-actions": "^2.6.5",
-    "redux-devtools-extension": "^2.13.8"
+    "redux-devtools-extension": "^2.13.8",
+    "redux-persist": "^6.0.0"
   }
 }
