@@ -1,5 +1,5 @@
 import {createActions} from 'redux-actions';
 
-export const {addProducts,} = createActions('ADD_PRODUCTS');
+export const {addProducts} = createActions('ADD_PRODUCTS');
 
 
