@@ -11,7 +11,7 @@ export default {
     basket: handleActions({
         [addToBasket]:(state, {payload = defaultPayload})=>{
             const newState={...state,productsID:[...state.productsID]};
-            newState.productsID.push(payload.productId);
+            newState.productsID.push(payload.productsID);
             newState.count+=1;
             newState.amount+=payload.priceValue;
             return newState
@@ -26,4 +26,5 @@ export default {
 
     },basketState)
 }
+
 
