@@ -1,6 +1,7 @@
 import {handleActions} from 'redux-actions';
+import cloneDeep from 'lodash.clonedeep';
 //action
-import {addProducts} from "../actions";
+import {addProducts, sortDesc, sortAsc} from "../actions";
 //state
 import {productsState} from "../constants/defaultState";
 
@@ -8,6 +9,26 @@ export default {
     products: handleActions({
         [addProducts]: (state, {payload = []}) => {
             return payload;
-        }
+        },
+        [sortDesc]: (state) => {
+
+            const newState=cloneDeep(state)
+            newState.sort((a, b) => {
+                if (a.price.value > b.price.value) return -1;
+                else return 1;
+            });
+            return newState
+
+        },
+        [sortAsc]: (state) => {
+
+            const newState=cloneDeep(state)
+            newState.sort((a, b) => {
+                if (a.price.value > b.price.value) return 1;
+                else return -1;
+            });
+            return newState
+
+        },
     }, productsState)
 }
